moved getRelativeDate into @core/utils/date, removed R dependancy
